+++ backend/server.js
@@ -1,6 +1,6 @@
 import http from 'http';
 import express from 'express';
-import * as db from './db_work.js';
+import * as db from './db.js';
 
 const server = http.createServer(async (req, res) => {
   //TODO: route endpoints
